Fixes bug when history gets too long from a command
When the history array got to look it would trim the array by 1.  If a
command was run and it generated multiple lines that exceeded the
maximum height of the canvas we’d get some strange behavior on the
current line.  The cursor would drop below canvas.  This cleans up the
canvas when exceeding height.